# specs/001-fullstack-todo-app/contracts/api-spec.yaml
openapi: 3.0.3
info:
  title: Phase II Full-Stack Todo Application API
  description: |
    REST API for multi-user todo application with JWT-based authentication and strict user isolation.

    **Security Model**:
    - All endpoints require JWT Bearer token in Authorization header
    - User identity derived exclusively from JWT payload, never from path parameters
    - Backend validates task ownership on every operation (403 Forbidden if violation)
    - Unauthenticated requests rejected with 401 Unauthorized

    **Scope**:
    - Phase II Basic Level features only (5 CRUD operations)
    - No priorities, tags, search, filtering, sorting, dates, AI, RBAC, uploads, or notifications

    **Reference**: @specs/001-fullstack-todo-app/spec.md (FR-027 through FR-036)
  version: 1.0.0
  contact:
    name: Phase II Development Team
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.todo-app.example.com
    description: Production server (placeholder)

security:
  - BearerAuth: []

paths:
  /api/{user_id}/tasks:
    get:
      summary: List all tasks for authenticated user
      description: |
        Retrieves all tasks belonging to the authenticated user. Tasks ordered by creation date (newest first).

        **Security Note**: The {user_id} path parameter is IGNORED. User identity derived from JWT token only.
        Backend validates that authenticated user matches expected user_id (403 if mismatch).
      operationId: listTasks
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID (ignored - user derived from JWT token for security)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
              examples:
                multipleTask:
                  summary: User with multiple tasks
                  value:
                    - id: 1
                      user_id: "550e8400-e29b-41d4-a716-446655440000"
                      title: "Buy groceries"
                      description: "Milk, eggs, bread"
                      completed: false
                      created_at: "2025-12-30T10:00:00Z"
                      updated_at: "2025-12-30T10:00:00Z"
                    - id: 2
                      user_id: "550e8400-e29b-41d4-a716-446655440000"
                      title: "Call dentist"
                      description: null
                      completed: true
                      created_at: "2025-12-29T15:30:00Z"
                      updated_at: "2025-12-30T09:00:00Z"
                emptyList:
                  summary: User with no tasks
                  value: []
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      security:
        - BearerAuth: []

    post:
      summary: Create new task
      description: |
        Creates a new task for the authenticated user. Task created with completed=false by default.

        **Security Note**: The {user_id} path parameter is IGNORED. Task automatically associated with
        authenticated user from JWT token.
      operationId: createTask
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID (ignored - user derived from JWT token)
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
            examples:
              withDescription:
                summary: Task with description
                value:
                  title: "Finish project report"
                  description: "Include Q4 metrics and recommendations"
              titleOnly:
                summary: Task without description
                value:
                  title: "Call mom"
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: 3
                user_id: "550e8400-e29b-41d4-a716-446655440000"
                title: "Finish project report"
                description: "Include Q4 metrics and recommendations"
                completed: false
                created_at: "2025-12-30T14:00:00Z"
                updated_at: "2025-12-30T14:00:00Z"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      security:
        - BearerAuth: []

  /api/{user_id}/tasks/{id}:
    get:
      summary: Get task details
      description: |
        Retrieves detailed information for a specific task. Backend verifies task belongs to authenticated user.
      operationId: getTask
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID (ignored - user derived from JWT token)
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: 1
                user_id: "550e8400-e29b-41d4-a716-446655440000"
                title: "Buy groceries"
                description: "Milk, eggs, bread"
                completed: false
                created_at: "2025-12-30T10:00:00Z"
                updated_at: "2025-12-30T10:00:00Z"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - BearerAuth: []

    put:
      summary: Update task title and description
      description: |
        Updates task title and/or description. Backend verifies task belongs to authenticated user.
        Completion status NOT updated by this endpoint (use PATCH /complete instead).
      operationId: updateTask
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID (ignored - user derived from JWT token)
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
            example:
              title: "Buy groceries and supplies"
              description: "Milk, eggs, bread, and cleaning supplies"
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: 1
                user_id: "550e8400-e29b-41d4-a716-446655440000"
                title: "Buy groceries and supplies"
                description: "Milk, eggs, bread, and cleaning supplies"
                completed: false
                created_at: "2025-12-30T10:00:00Z"
                updated_at: "2025-12-30T15:00:00Z"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - BearerAuth: []

    delete:
      summary: Delete task permanently
      description: |
        Permanently deletes a task. Backend verifies task belongs to authenticated user. No soft delete in Phase II.
      operationId: deleteTask
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID (ignored - user derived from JWT token)
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            minimum: 1
      responses:
        '204':
          description: Task deleted successfully (no content)
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - BearerAuth: []

  /api/{user_id}/tasks/{id}/complete:
    patch:
      summary: Toggle task completion status
      description: |
        Marks task as complete or incomplete. Backend verifies task belongs to authenticated user.
      operationId: toggleTaskCompletion
      tags:
        - Tasks
      parameters:
        - name: user_id
          in: path
          required: true
          description: User ID (ignored - user derived from JWT token)
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          description: Task ID
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCompletionToggle'
            examples:
              markComplete:
                summary: Mark task as complete
                value:
                  completed: true
              markIncomplete:
                summary: Mark task as incomplete
                value:
                  completed: false
      responses:
        '200':
          description: Task completion status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
              example:
                id: 1
                user_id: "550e8400-e29b-41d4-a716-446655440000"
                title: "Buy groceries"
                description: "Milk, eggs, bread"
                completed: true
                created_at: "2025-12-30T10:00:00Z"
                updated_at: "2025-12-30T16:00:00Z"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token issued by Better Auth upon successful authentication.

        **Token Format**: Authorization: Bearer <token>

        **Token Payload**:
        - user_id: string (UUID)
        - email: string
        - iat: number (issued at timestamp)
        - exp: number (expiration timestamp)

        **Security Requirements**:
        - Backend MUST verify JWT signature using BETTER_AUTH_SECRET
        - Backend MUST validate token expiration
        - Backend MUST extract user_id from payload (never from request)
        - Invalid/expired/missing tokens → 401 Unauthorized

  schemas:
    Task:
      type: object
      description: Complete task object with all fields
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique task identifier
          example: 1
        user_id:
          type: string
          format: uuid
          description: Owner of this task (UUID)
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Task title (required, 1-255 characters)
          example: "Buy groceries"
        description:
          type: string
          nullable: true
          maxLength: 2000
          description: Optional task description (0-2000 characters)
          example: "Milk, eggs, bread"
        completed:
          type: boolean
          description: Completion status (false = incomplete, true = complete)
          example: false
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp (ISO 8601)
          example: "2025-12-30T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last modification timestamp (ISO 8601, auto-updated)
          example: "2025-12-30T10:00:00Z"

    TaskCreate:
      type: object
      description: Request body for creating new task
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Task title (required, 1-255 characters)
          example: "Finish project report"
        description:
          type: string
          nullable: true
          maxLength: 2000
          description: Optional task description (0-2000 characters)
          example: "Include Q4 metrics and recommendations"

    TaskUpdate:
      type: object
      description: Request body for updating task title/description
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Updated task title (required, 1-255 characters)
          example: "Buy groceries and supplies"
        description:
          type: string
          nullable: true
          maxLength: 2000
          description: Updated task description (0-2000 characters, optional)
          example: "Milk, eggs, bread, and cleaning supplies"

    TaskCompletionToggle:
      type: object
      description: Request body for toggling task completion status
      required:
        - completed
      properties:
        completed:
          type: boolean
          description: New completion status (true = complete, false = incomplete)
          example: true

    ErrorResponse:
      type: object
      description: Standard error response format
      required:
        - detail
      properties:
        detail:
          type: string
          description: Human-readable error message
          example: "Task not found"

  responses:
    UnauthorizedError:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              summary: Missing Authorization header
              value:
                detail: "Not authenticated"
            invalidToken:
              summary: Invalid JWT signature
              value:
                detail: "Could not validate credentials"
            expiredToken:
              summary: Expired JWT token
              value:
                detail: "Token has expired"

    ForbiddenError:
      description: Valid token but user lacks permission to access resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            userMismatch:
              summary: User_id mismatch (path vs JWT)
              value:
                detail: "Access denied: user ID mismatch"
            taskOwnershipViolation:
              summary: Task belongs to different user
              value:
                detail: "Access denied: task belongs to another user"

    NotFoundError:
      description: Requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Task not found"

    ValidationError:
      description: Request validation failed (invalid input)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            emptyTitle:
              summary: Empty task title
              value:
                detail: "Title cannot be empty"
            titleTooLong:
              summary: Title exceeds 255 characters
              value:
                detail: "Title must be 255 characters or less"
            descriptionTooLong:
              summary: Description exceeds 2000 characters
              value:
                detail: "Description must be 2000 characters or less"

tags:
  - name: Tasks
    description: Task CRUD operations (create, read, update, delete, mark complete)